Emulation is great, released Crtpl3 as 0.9, project updates
Look, up there in the cloud! Is it a plane? Is it a bir-no it's a gigantic commit because I have no idea how version control works.
As an aside, there are now updates in the right sidebar - helpful to find my latest stuff and finally puts to use the modify dates I've been preparing for a while.
Still need to polish crtpl3 - low-width environments, settings dialog, plotting support, more tools.